+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `CSVTableReader` and `CSVTableWriter` constructors add `readRecords` parameter.
+- `CSVDataHelper.GetCSVRowArray` and `CSVDataHelper.GetCSVRowList` add `rowCount` parameter.
+- `CSVDataHelper` add `HeaderInfoRowCount` constant.
 
 ## [1.2.0] - 2020-09-18
 ### Added